Adjust the script to run outside of an Azure Pipeline
For developing the script, it is crucial that it works in a local
checkout.

So let's handle the situation where `last.modified` is not interpolated
and pretend that it is a known-past value (rather than using a local
script with the name `last.modified` that needs to be in the `PATH` and
that is expected to output such a known-past Last-Modified: value).

While at it, only download the tar file if it has not been downloaded
already (and is new enough).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>